New extended 6" board with 5W resistors.
Still needs the transistor pad sizes reduced. Decided not to add a
jumper for pull down resistor for current measurement sensor. Instead
this resistor will be external to the board. It will work with the DHAB
S/14. Did not change pad size on 0603. Large through hole fuse, zener
and schottky diode for protection.